requirements.txt
alabaster==0.7.13
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
billiard==4.1.0
black==23.10.1
boto3==1.28.77
botocore==1.31.77
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.7
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.26.5
exceptiongroup==1.1.3
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kombu==5.3.2
livereload==2.6.3
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg-c==3.1.12
Pygments==2.16.1
pytest==7.4.3
pytest-django==4.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
s3transfer==0.7.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.9
gunicorn==20.1.0
config==0.5.1